agents/retriever_agent/requirements.txt
fastapi
uvicorn
requests
faiss-cpu
sentence-transformers
numpy
